requirements.txt
freetype-py==2.2.0
gif2numpy==1.3
kaitaistruct==0.9
MouseInfo==0.1.3
mss==6.1.0
numpy==1.19.4
numpy2gif==1.0
opencv-python==4.4.0.46
Pillow==8.0.1
PyAutoGUI==0.9.52
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
PyTweening==1.0.3
